文章中转站/pyproject.toml
[project]
name = "content_hub"
version = "0.1.0"
description = "service-first content hub extracted from AIWriteX"
authors = [{ name = "iniwap", email = "iniwaper@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
  "PyYAML==6.0.1",
  "beautifulsoup4==4.13.4",
  "fastapi>=0.117.1",
  "jinja2>=3.1.0",
  "markdown==3.8.2",
  "pydantic",
  "requests==2.32.3",
  "tomlkit==0.13.2",
  "uvicorn>=0.30.0",
]

[project.scripts]
content_hub_api = "content_hub.interfaces.api.main:create_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
